Moved sample.py content to montecarlo.py
